move fe_testutils + sec_ord_ell_bvp into lf/fe/test
