+++ lib/lf/fe/fe.h
@@ -15,13 +15,13 @@
  * @copyright MIT License
  */
 
-#include "fe_testutils.h"
+#include "fe_tools.h"
 #include "lin_fe.h"
 #include "loc_comp_norms.h"
 #include "mesh_function_constant.h"
 #include "mesh_function_global.h"
 #include "mesh_function_traits.h"
-#include "sec_ord_ell_bvp.h"
+#include "uniform_scalar_finite_element_space.h"
 
 namespace lf::fe {}  // namespace lf::fe
