+++ lib/lf/fe/test/bvp_fe_test.cc
@@ -13,13 +13,13 @@
  */
 
 #include <gtest/gtest.h>
+#include <lf/fe/fe.h>
 #include <iostream>
-#include "lf/fe/fe_testutils.h"
-#include "lf/fe/lin_fe.h"
-#include "lf/fe/sec_ord_ell_bvp.h"
+#include "fe_testutils.h"
 
 #include <lf/mesh/utils/utils.h>
 #include "lf/mesh/test_utils/test_meshes.h"
+#include "sec_ord_ell_bvp.h"
 
 #define REFLEV 5
